* further repo restructuring, deleting extras, made more importable * ec2stop ready for deployment * made more active use of account id * red and green buttons now fully implemented * vault fully implemented * work started on find by content
