+++ security_scripts/cfg/defaults.cfg
@@ -2,25 +2,33 @@
 profile=scimma-uiuc-aws-admin
 role=scimma_power_user
 vaultdir=~/.vault
+loglevel=INFO
 
 # info tools
 [TAG_REPORT]
-loglevel=NORMAL
+# loglevel unused
+loglevel=INFO
 dbfile=:memory:
 [FIND_BY_CONTENT]
+# loglevel unused
 loglevel=INFO
 [DOWNLOAD]
+# loglevel unused
 loglevel=INFO
 bucket=s3://scimma-processes/Scimma-event-trail
 accountid=585193511743
 [S3_REPORT]
 profile=scimma-uiuc-aws-admin
+# loglevel unused
 loglevel=NORMAL
 
 # control tools
 [BUTTONS]
+# loglevel unused
 loglevel=DEBUG
 [AUDIT]
+# loglevel unused
 loglevel=WARNING
 [TESTS]
+# loglevel unused
 LOGLEVEL=DEBUG
