+++ security_scripts/information/.gitignore
@@ -1,0 +1,3 @@
+__pycache__
+*~
+\#*\#
