+++ security_scripts/information/README.vault.txt
@@ -1,0 +1,10 @@
+Vault.py is a command to download the SCiMMA
+Cloudtrail data into a directory under your login
+area, $HOME/.trailscraper.
+
+Downloads are incremental -- previous downloads are not
+re-fetched or deleted.
+
+Other tools (notably find_by_content.py) based
+on trailscraper use the downloaded data for analysis.
+
