--- security_scripts/information/bin/.gitignore
@@ -1,3 +1,0 @@
-__pycache__
-*~
-\#*\#
