--- security_scripts/information/bin/README.md
@@ -1,1 +1,0 @@
-Placeholder
