--- security_scripts/information/bin/vault/.gitignore
@@ -1,1 +1,0 @@
-Scimma-event-trail
